+++ tests/dummy/app/routes/components-examples/flexberry-objectlistview/on-edit-form/suggestion.js
@@ -1,0 +1,32 @@
+import EditFormRoute from 'ember-flexberry/routes/edit-form';
+
+export default EditFormRoute.extend({
+  /**
+    Name of model projection to be used as record's properties limitation.
+
+    @property modelProjection
+    @type String
+    @default 'SuggestionE'
+   */
+  modelProjection: 'SuggestionE',
+
+  /**
+  For default userSetting use empty name ('').
+  <componentName> may contain any of properties: colsOrder, sorting, colsWidth or being empty.
+
+  @property developerUserSettings
+  @type Object
+  @default {}
+  */
+
+  developerUserSettings: { listOnEditform: { } },
+
+  /**
+    Name of model to be used as form's record type.
+
+    @property modelName
+    @type String
+    @default 'ember-flexberry-dummy-suggestion'
+   */
+  modelName: 'ember-flexberry-dummy-suggestion',
+});